improved ai-chatbot performance
Moved ai functions to src/services/ai_chatbot.py, modified server start to call initialize_ai_system(), which loads dataset and computes vectores. Eventually will have to set it up so vectors are already calculated for each game ahead of time, and are just loaded into the model on start
