+++ gamelib-backend/src/api/users.py
@@ -1,155 +1,37 @@
-from typing import Optional
+from typing import Optional, Dict, Any
 from fastapi import APIRouter, HTTPException
-from src.db.supabase_client import supabase
+from src.db.repositories.users_db import UsersRepository
 from src.schemas.user_schema import User, fetch_steam_profile
 from postgrest.exceptions import APIError
-import asyncio
 
 router = APIRouter()
 
 
-
 @router.post("/users/", response_model=User)
 async def create_user(user: User):
+    """Create a new user"""
     try:
         # Validate steam_id
         if user.steam_id <= 0 or user.steam_id > 999999999999999999:
             raise HTTPException(status_code=400, detail="Invalid Steam ID")
         
-        # Create games_array if games data is provided
-        insert_data = {
-            "steam_id": user.steam_id, 
-            "data": user.data, 
-            "login_count": user.login_count
-        }
+        # Use repository method with automatic games_array generation
+        created_user = UsersRepository.create_with_games(
+            steam_id=user.steam_id,
+            data=user.data or {},
+            games=user.games or {},
+            login_count=user.login_count
+        )
         
-        # Add games and games_array if games exist
-        if user.games is not None:
-            insert_data["games"] = user.games
-            # Sort game IDs by playtime
-            user_games = user.games  # Type narrowing
-            games_array = sorted(
-                user_games.keys(),
-                key=lambda gid: user_games[gid].get('playtime_forever', 0) if isinstance(user_games[gid], dict) else 0,
-                reverse=True
-            )
-            insert_data["games_array"] = games_array
-        
-        response = supabase.table("users").insert(insert_data).execute()
-        if not response.data:
+        if not created_user:
             raise HTTPException(status_code=400, detail="Failed to create user")
-        return User.model_validate(response.data[0])
+        
+        return created_user
     except APIError as e:
         raise HTTPException(status_code=400, detail=e.message)
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
 
-async def user_login(steam_id: int):
-    print(f"User login attempt for steam_id: {steam_id}")
-    # Check if user already exists
-    response = supabase.table('users').select('*').eq('steam_id', steam_id).execute()
-    
-    if response.data:
-        # User already exists, update Steam data and increment login_count
-        user = await update_user_data(steam_id)
-        return user.model_dump() if user else None
-    else:
-        print("Creating new user")
-        
-        # Fetch Steam profile (includes both games and player profile)
-        profile_data = await fetch_steam_profile(steam_id)
-        
-        if not profile_data or 'profile' not in profile_data:
-            raise Exception("Could not fetch Steam player profile")
-        
-        # Extract player profile and games from response
-        player_profile = profile_data['profile']
-        games = profile_data.get('games', {})
-        games_array = profile_data.get('games_array', [])
-        
-        # Store the player profile data (which includes personaname)
-        user_create = User(
-            steam_id=steam_id, 
-            data=player_profile, 
-            games=games,
-            games_array=games_array,
-            login_count=1
-        )
-        print(f"Creating user with data: {user_create}")
-        created = await create_user(user_create)
-        
-        # Fetch and return the created user
-        response = supabase.table('users').select('*').eq('steam_id', steam_id).execute()
-        if response.data:
-            print(f"Created user: {response.data[0]}")
-            return response.data[0]
-        return None
-
-async def get_user_data(steam_id: int):
-    """Retrieve existing user data from database without updating"""
-    try:
-        response = supabase.table('users').select('*').eq('steam_id', steam_id).execute()
-        if response.data and len(response.data) > 0:
-            return User.model_validate(response.data[0])
-        return None
-    except Exception as e:
-        print(f"Error in get_user_data: {str(e)}")
-        return None
-
-async def update_user_data(steam_id: int):
-    try:
-        # Fetch fresh Steam profile (includes both games and player profile)
-        print(f"Fetching Steam profile for steam_id: {steam_id}")
-        profile_data = await fetch_steam_profile(steam_id)
-        
-        if not profile_data or 'profile' not in profile_data:
-            raise Exception("Could not fetch Steam player profile")
-        
-        # Extract player profile and games from response
-        player_profile = profile_data['profile']
-        games = profile_data.get('games', {})
-        games_array = profile_data.get('games_array', [])
-        
-        print(f"Fetched Steam player profile: {player_profile}")
-        
-        # Get current user to increment login_count
-        current_user_response = supabase.table('users').select('login_count').eq('steam_id', steam_id).execute()
-        
-        current_login_count = 1  # Default for new users
-        if current_user_response.data and len(current_user_response.data) > 0:
-            existing_count = current_user_response.data[0].get('login_count', 0)
-            current_login_count = existing_count + 1
-            print(f"Found existing login_count: {existing_count}, incrementing to: {current_login_count}")
-        else:
-            print(f"No existing user found, using default login_count: {current_login_count}")
-        
-        # Update database with fresh player profile data and incremented login_count
-        update_payload = {
-            'data': player_profile,
-            'login_count': current_login_count,
-            'games': games,
-            'games_array': games_array
-        }
-        
-        print(f"Update payload: {update_payload}")
-        
-        response = supabase.table('users').update(update_payload).eq('steam_id', steam_id).execute()
-        print(f"Updated login_count to: {current_login_count}")
-        
-        if response.data and len(response.data) > 0:
-            print(f"Returning updated user: {response.data[0]}")
-            return User.model_validate(response.data[0])
-        else:
-            print("No data returned from database update")
-            # If update didn't return data, fetch the user record
-            get_response = supabase.table('users').select('*').eq('steam_id', steam_id).execute()
-            if get_response.data and len(get_response.data) > 0:
-                print(f"Fetched user after update: {get_response.data[0]}")
-                return User.model_validate(get_response.data[0])
-            return None
-    except Exception as e:
-        print(f"Error in update_user_data: {str(e)}")
-        return None
 
 @router.get("/users/{steam_id}", response_model=User)
 async def get_user(steam_id: int, refresh: bool = False):
@@ -164,11 +46,15 @@ async def get_user(steam_id: int, refresh: bool = False):
             raise HTTPException(status_code=400, detail="Invalid Steam ID")
         
         if refresh:
-            # Fetch fresh Steam data and update database
-            user_data = await update_user_data(steam_id)
+            # Fetch fresh Steam data
+            profile_data = await fetch_steam_profile(steam_id)
+            if not profile_data or 'profile' not in profile_data:
+                raise HTTPException(status_code=400, detail="Could not fetch Steam profile")
+            
+            user_data = await UsersRepository.refresh_user_steam_data(steam_id, profile_data)
         else:
-            # Just get existing data from database
-            user_data = await get_user_data(steam_id)
+            # Get existing data from database
+            user_data = UsersRepository.find_by_steam_id(steam_id)
         
         if not user_data:
             raise HTTPException(status_code=404, detail="User not found")
@@ -176,10 +62,10 @@ async def get_user(steam_id: int, refresh: bool = False):
         return user_data
     except APIError as e:
         raise HTTPException(status_code=400, detail=e.message)
+    except HTTPException:
+        raise
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
-
-
 
 
 @router.put("/users/{steam_id}", response_model=User)
@@ -193,29 +79,32 @@ async def update_user(steam_id: int, user: Optional[User] = None, refresh_steam:
     try:
         if user and user.data:
             # Update with provided user data
-            update_data: dict = {"data": user.data}
-            if user.login_count is not None:
-                update_data["login_count"] = user.login_count
-            
-            # Add games and games_array if games exist
-            if user.games is not None:
-                update_data["games"] = user.games
-                # Sort game IDs by playtime
-                user_games = user.games  # Type narrowing
-                games_array = sorted(
-                    user_games.keys(),
-                    key=lambda gid: user_games[gid].get('playtime_forever', 0) if isinstance(user_games[gid], dict) else 0,
-                    reverse=True
+            if user.games:
+                # Use repository method with automatic games_array generation
+                updated_user = UsersRepository.update_with_games(
+                    steam_id=steam_id,
+                    data=user.data,
+                    games=user.games,
+                    increment_login=False
                 )
-                update_data["games_array"] = games_array
+            else:
+                # Simple update without games
+                update_data: Dict[str, Any] = {"data": user.data}
+                if user.login_count is not None:
+                    update_data["login_count"] = user.login_count
+                updated_user = UsersRepository.update(steam_id, update_data)
             
-            response = supabase.table("users").update(update_data).eq("steam_id", steam_id).execute()
-            if not response.data:
+            if not updated_user:
                 raise HTTPException(status_code=400, detail="Failed to update user")
-            return User.model_validate(response.data[0])
+            return updated_user
+            
         elif refresh_steam:
-            # Fetch fresh Steam data and update (this will increment login_count)
-            user_data = await update_user_data(steam_id)
+            # Fetch fresh Steam data
+            profile_data = await fetch_steam_profile(steam_id)
+            if not profile_data or 'profile' not in profile_data:
+                raise HTTPException(status_code=400, detail="Could not fetch Steam profile")
+            
+            user_data = await UsersRepository.refresh_user_steam_data(steam_id, profile_data)
             if not user_data:
                 raise HTTPException(status_code=400, detail="Failed to update user with Steam data")
             return user_data
@@ -223,10 +112,10 @@ async def update_user(steam_id: int, user: Optional[User] = None, refresh_steam:
             raise HTTPException(status_code=400, detail="Either provide user data or set refresh_steam=True")
     except APIError as e:
         raise HTTPException(status_code=400, detail=e.message)
+    except HTTPException:
+        raise
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
-
-
 
 
 @router.post("/users/{steam_id}/refresh", response_model=User)
@@ -235,27 +124,35 @@ async def refresh_user_steam_data(steam_id: int):
     Refresh user data by fetching fresh data from Steam API and updating database.
     """
     try:
-        user_data = await update_user_data(steam_id)
+        profile_data = await fetch_steam_profile(steam_id)
+        if not profile_data or 'profile' not in profile_data:
+            raise HTTPException(status_code=400, detail="Could not fetch Steam profile")
+        
+        user_data = await UsersRepository.refresh_user_steam_data(steam_id, profile_data)
         if not user_data:
             raise HTTPException(status_code=404, detail="User not found or failed to refresh Steam data")
         return user_data
     except APIError as e:
         raise HTTPException(status_code=400, detail=e.message)
+    except HTTPException:
+        raise
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
 
 
 @router.delete("/users/{steam_id}")
 async def delete_user(steam_id: int):
+    """Delete a user by steam_id"""
     try:
-        response = supabase.table("users").delete().eq("steam_id", steam_id).execute()
-        if not response.data:
+        success = UsersRepository.delete(steam_id)
+        if not success:
             raise HTTPException(status_code=400, detail="Failed to delete user")
         return {"detail": "User deleted successfully"}
     except APIError as e:
         raise HTTPException(status_code=400, detail=e.message)
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
+
 
 @router.get("/users/{steam_id}/name")
 async def get_user_name(steam_id: int):
@@ -264,36 +161,31 @@ async def get_user_name(steam_id: int):
     If user doesn't exist in database, triggers login process to create them
     """
     try:
-        
         # First try to get from database
-        user_data = await get_user_data(steam_id)
+        user_data = UsersRepository.find_by_steam_id(steam_id)
         
         if user_data:
             # User exists in database, get name from stored data
-            steam_data = user_data.data or {}
-            player_name = steam_data.get('personaname', 'Unknown Player')
+            player_name = user_data.persona_name
             print(f"Found user in database, player_name: {player_name}")
             return {"name": player_name}
         else:
-            # User doesn't exist in database, trigger login process to create them
+            # User doesn't exist, fetch from Steam and create
+            print(f"User {steam_id} not found, creating...")
+            profile_data = await fetch_steam_profile(steam_id)
             
-            try:
-                login_result = await user_login(steam_id)
-                print(f"Login result: {login_result}")
-                
-                if login_result:
-                    # User was successfully created/updated, get name from database result
-                    steam_data = login_result.get('data', {})
-                    player_name = steam_data.get('personaname', 'Unknown Player')
-                    print(f"Created user successfully, player_name: {player_name}")
-                    return {"name": player_name}
-                else:
-                    print("Login result was None")
-                    raise HTTPException(status_code=500, detail="Failed to create user - login returned None")
-                    
-            except Exception as login_error:
-                print(f"Error during user_login: {login_error}")
-                raise HTTPException(status_code=500, detail=f"Failed to create user: {str(login_error)}")
+            if not profile_data or 'profile' not in profile_data:
+                raise HTTPException(status_code=400, detail="Could not fetch Steam profile")
+            
+            # Create user via login
+            login_result = await UsersRepository.user_login(steam_id, profile_data)
+            
+            if login_result:
+                player_name = login_result.get('data', {}).get('personaname', 'Unknown Player')
+                print(f"Created user successfully, player_name: {player_name}")
+                return {"name": player_name}
+            else:
+                raise HTTPException(status_code=500, detail="Failed to create user")
                 
     except HTTPException:
         raise
